Apply styling
