--- src/index.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Leaderboard</title>
-  </head>
-  <body>
-    <main id="container" class="container">
-      <section class="scores-container">
-        <h1 class="leaderboard-title">Leaderboard</h1>
-        <div class="leaderboard-container">
-          <div class="recent-scores">
-            <div class="title">
-              <h2>Recent Scores</h2>
-              <button class="refresh-btn clickable" type="button">
-                Refresh
-              </button>
-            </div>
-            <ul class="list">
-              <li class="score-item">
-                <span>Name: </span>
-                <span>100</span>
-              </li>
-              <li class="score-item">
-                <span>Name: </span>
-                <span>20</span>
-              </li>
-              <li class="score-item">
-                <span>Name: </span>
-                <span>50</span>
-              </li>
-              <li class="score-item">
-                <span>Name: </span>
-                <span>78</span>
-              </li>
-              <li class="score-item">
-                <span>Name: </span>
-                <span>125</span>
-              </li>
-              <li class="score-item">
-                <span>Name: </span>
-                <span>77</span>
-              </li>
-              <li class="score-item">
-                <span>Name: </span>
-                <span>42</span>
-              </li>
-            </ul>
-          </div>
-          <form class="form">
-            <h2>Add your score</h2>
-            <input type="text" class="input-name" placeholder="Your Name" />
-            <input type="text" class="input-score" placeholder="Your Score" />
-            <button class="submit-btn clickable" type="button">Submit</button>
-          </form>
-        </div>
-      </section>
-    </main>
-  </body>
-</html>
